+++ django_tiptap_editor/fields/__init__.py
@@ -1,0 +1,7 @@
+"""Model fields."""
+
+from __future__ import annotations
+
+from django_tiptap_editor.fields.tiptap_json_field import TipTapJSONField
+
+__all__ = ["TipTapJSONField"]