+++ django_tiptap_editor/forms/__init__.py
@@ -3,5 +3,6 @@
 from __future__ import annotations
 
 from django_tiptap_editor.forms.fields import TipTapFormField
+from django_tiptap_editor.forms.json_field import TipTapJSONFormField
 
-__all__ = ["TipTapFormField"]
+__all__ = ["TipTapFormField", "TipTapJSONFormField"]